+++ src/ngfmdmasync.h
@@ -17,8 +17,8 @@ class ngfmdmasync:public bufferdma,public clkgpio,public pwmgpio,public pcmgpio
 	void SetDmaAlgo();
 	
 	void SetPhase(bool inversed);
-	void SetFrequencySample(uint32_t Index,double Frequency);
-	void SetFrequencySamples(double *sample,size_t Size);
+	void SetFrequencySample(uint32_t Index,float Frequency);
+	void SetFrequencySamples(float *sample,size_t Size);
 }; 
 
 #endif
